+++ PSModule/M365Documentation/Functions/Get-M365Doc.ps1
@@ -20,9 +20,7 @@ Function Get-M365Doc(){
     The Script is using the MSAL.PS Module and requires you to run Connect-Documentation. Therefore you have to install them first.
 
     .PARAMETER Components
-        Path including filename where the documentation should be created. The filename has to end with .docx.
-        Note:
-        If there is already a file present, the documentation will be added at the end of the existing document.
+        The components for which the available sections should be returned.
 
     .PARAMETER ExcludeSections
         The specified sections are excluded from the collection process. The following sections are available:
@@ -124,6 +122,12 @@ Function Get-M365Doc(){
     Retrieves data from backup file. created with Output-M365Doc
     
     $doc = Get-M365Doc -BackupFile c:\temp\backup.json
+
+    .Example Select Sections in UI with Out-GridView allowing selection of sections
+    $Selection = Get-M365DocValidSection | Out-GridView -OutputMode Multiple
+    $Sections = $Selection | Select-Object -ExpandProperty SectionName
+    $Components = $Selection | Select-Object -ExpandProperty Component -Unique
+    $doc = Get-M365Doc -Components $Components -IncludeSections $Sections
 
     .NOTES
     Author: Thomas Kurth/baseVISION